[NB] Fix IO packing...
Fix IO packing
==============

Issue
-----

Prior to this commit, although the input/output slot-block-keys were
packed, such that each IO tile was filled contiguously starting with the
first position in the tile, the resulting block positions were not
synchronized back to the block-slot-keys array.

Fix
---

This commit adds a synchronization from the packed slot-block-keys to
the block-slot-keys during the `CAMIP::finalize` method.